Migrate gitter tests to pytest
