+++ opsdroid/connector/gitter/tests/responses/gitter_bot_message.json
@@ -1,0 +1,7 @@
+{
+    "text": "hi",
+    "fromUser": {
+        "username": "bot",
+        "id": "12345"
+    }
+}
