+++ opsdroid/connector/gitter/tests/responses/gitter_connect.json
@@ -1,0 +1,5 @@
+{
+    "login": "opsdroid",
+    "id": "12345",
+    "username": "bot"
+}
