+++ opsdroid/connector/gitter/tests/responses/gitter_message.json
@@ -1,0 +1,7 @@
+{
+    "text": "hi",
+    "fromUser": {
+        "username": "not a bot",
+        "id": "67890"
+    }
+}
